+++ src/config-load.js
@@ -1,10 +1,14 @@
 /*
-db    db  .8888.  dP     888888b 8888ba   .8888.     d8b   db  888888b d8888P
-88    88 d8'  `8b 88     88      88  `8b d8'  `8b    88V8  88  88        88
-Y8    8P 88    88 88     88aaa   88aa8P' 88    88    88 V8 88 a88aaa     88
-`8b  d8' 88    88 88     88      88  `8b 88    88    88  V888  88        88
- `8bd8'  Y8.  .8P 88     88      88  .88 Y8.  .8P dP 88   V88  88        88
-   YP     `888P'  88888P 888888P 888888'  `888P'  88 VP    8P  888888P   dP
+################################################################################
+#                                                                              #
+# db    db  .8888.  dP     888888b 8888ba   .8888.     d8b   db 888888b d8888P #
+# 88    88 d8'  `8b 88     88      88  `8b d8'  `8b    88V8  88 88        88   #
+# Y8    8P 88    88 88    a88aaa   88aa8P' 88    88    88 V8 88 88aaa     88   #
+# `8b  d8' 88    88 88     88      88  `8b 88    88    88  V888 88        88   #
+#  `8bd8'  Y8.  .8P 88     88      88  .88 Y8.  .8P dP 88   V88 88        88   #
+#    YP     `888P'  88888P 888888P 888888'  `888P'  88 VP    8P 888888P   dP   #
+#                                                                              #
+################################################################################
 
 ExpressJS for volebo.net
 
@@ -18,11 +22,11 @@ the Free Software Foundation, either version 3 of the License, or
 
 This program is distributed in the hope that it will be useful,
 but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
 GNU General Public License for more details.
 
 You should have received a copy of the GNU General Public License
-along with this program.  If not, see <http://www.gnu.org/licenses/>.
+along with this program. If not, see <http://www.gnu.org/licenses/>.
 */
 
 'use strict'
@@ -65,6 +69,9 @@ class Config {
 function loadConfig(configPath, overrideOptions) {
 	const defs = {
 		'volebo': {
+			// information from package.json (taken from main project)
+			//'package': undefined,
+
 			'server': {
 				'host': '127.0.0.1',
 				'port': 3000,
@@ -76,6 +83,10 @@ function loadConfig(configPath, overrideOptions) {
 
 				// TODO: move from this section!
 				'staticPath': 'public',
+			},
+
+			'sentry': {
+				'enabled': false,
 			},
 
 			'flash': {